Project_Epic/requirements.txt
# Project Epic - Python Dependencies

# AI Models
anthropic>=0.18.0
google-generativeai>=0.3.0

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Optional (for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
